Finished debugging AJAX GET call for grocery list
When I click "SHOP", I want to retrieve all items where
quantity = false. This means that I no longer have the item
and it needs to be added to my grocery list.

In the future, I will need to submit a PATCH to update the
quantity value and then retrieve it using this GET request.