Developed framework for handlebars (food list and grocery list)
When I click "FOOD", I want to display my existing food item names
and the number of days left to expiration.

When I click "SHOP", I want to display my grocery list (food name
and name of the store).

A very ugly list is now displayed on the page. The number of days
until expiration is not populating yet because I have yet to add that
variable to the model.