Added preliminary code for PATCH and DELETE AJAX calls
Modified handlebars to try and incorporate specific ids.
The back-end for both PATCH and DELETE are working.
I can update the value of "quantity" from true to false.
I can remove a row.

However, I'm having a BIG issue with handlebars. When I click
on either the button to delete the item or move it to the grocery
list, nothing happens. I think the back-end plumbing is in place
for this AJAX call to be successful, but I am not yet able to
execute on the front-end.