Added "remove-item" class to grocery-list handlebars
Once a user has bought the item, they can click the X and
delete the item from the database.

I will need a smooth way to navigate to the add form. The user
will effectively need to create a new item (with a new expiration)
date in order to add the item back into the list.